bank/requirements.txt
The operating system on which the project was developed:
    Windows 11 Home

REQUIREMENTS
    djangorestframework 3.13.1
    Django 4.1
    djoser 2.1.0
    python 3.10

PIP LIST:
    Package                       Version
    ----------------------------- ---------
    asgiref                       3.5.2
    certifi                       2022.6.15
    cffi                          1.15.1
    charset-normalizer            2.1.0
    coreapi                       2.3.3
    coreschema                    0.0.4
    cryptography                  37.0.4
    defusedxml                    0.7.1
    Django                        4.1
    django-rest-framework         0.1.0
    django-templated-mail         1.1.1
    djangorestframework           3.13.1
    djangorestframework-simplejwt 4.8.0
    djoser                        2.1.0
    idna                          3.3
    itypes                        1.2.0
    Jinja2                        3.1.2
    MarkupSafe                    2.1.1
    oauthlib                      3.2.0
    pip                           22.2.1
    pycparser                     2.21
    PyJWT                         2.4.0
    python3-openid                3.2.0
    pytz                          2022.2.1
    requests                      2.28.1
    requests-oauthlib             1.3.1
    setuptools                    63.2.0
    six                           1.16.0
    social-auth-app-django        4.0.0
    social-auth-core              4.3.0
    sqlparse                      0.4.2
    tzdata                        2022.2
    uritemplate                   4.1.1
    urllib3                       1.26.11

COMANDS IN TERMINAL:
    pip install djangorestframework
    pip install djoser